+++ extensions/supernotes/src/components/CommonCardActions.tsx
@@ -25,10 +25,25 @@ const CommonCardActions = ({ card, removeFromList }: CommonCardActionsProps) => 
 
   return (
     <>
-      <Action.OpenInBrowser url={`${SUPERNOTES_VIEW_URL}/${card.data.id}/`}/>
-      <Action.CopyToClipboard title="Copy Markdown" icon={Icon.Clipboard} content={card.data.markup} shortcut={{modifiers: ["cmd"], key: "."}}/>
-      <Action.CopyToClipboard title="Copy HTML" icon={Icon.Clipboard} content={card.data.html} shortcut={{modifiers: ["cmd", "shift"], key: "."}}/>
-      <Action title="Junk Card" icon={Icon.Trash} onAction={() => junk(card.data.id)} shortcut={{modifiers: ["ctrl"], key: "x"}}/>
+      <Action.OpenInBrowser url={`${SUPERNOTES_VIEW_URL}/${card.data.id}/`} />
+      <Action.CopyToClipboard
+        title="Copy Markdown"
+        icon={Icon.Clipboard}
+        content={card.data.markup}
+        shortcut={{ modifiers: ["cmd"], key: "." }}
+      />
+      <Action.CopyToClipboard
+        title="Copy HTML"
+        icon={Icon.Clipboard}
+        content={card.data.html}
+        shortcut={{ modifiers: ["cmd", "shift"], key: "." }}
+      />
+      <Action
+        title="Junk Card"
+        icon={Icon.Trash}
+        onAction={() => junk(card.data.id)}
+        shortcut={{ modifiers: ["ctrl"], key: "x" }}
+      />
       {found && <Action title="Remove from Recently Viewed" icon={Icon.XmarkCircle} onAction={remove} />}
     </>
   );