+++ extensions/supernotes/CHANGELOG.md
@@ -1,10 +1,11 @@
 # Supernotes Changelog
 
-## [Open Note in Browser] - 2022-07-15
+## [New Browser Shorcuts, Improved Creation Form] - 2022-07-15
 
 - Add shortcuts for home, daily, thoughts, inbox, and junk views
 - Allow open cards in the browser from the search and create commands
-- Improve submit action label in `Create Card` form
+- Default to showing an HUD on new card creation, add shortcut to open the preview
+- add a confirmation alert on card deletion
 
 ## [Initial Version] - 2022-07-15